+++ KalikoCMS.Engine/Data/Synchronizer.cs
@@ -70,6 +70,7 @@ namespace KalikoCMS.Data {
                     var pageType = Mapper.Map<PageTypeEntity, PageType>(pageTypeEntity);
                     pageType.Type = type;
                     pageType.AllowedTypes = attribute.AllowedTypes;
+                    pageType.CanCreate = attribute.CanCreate;
                     pageType.PreviewImage = attribute.PreviewImage;
                     pageType.Instance = (CmsPage)Activator.CreateInstance(type);